+++ backend/package.json
@@ -44,6 +44,8 @@
     "@babel/preset-env": "^7.0.0",
     "@babel/preset-typescript": "^7.26.0",
     "@babel/register": "^7.0.0",
+    "@types/cors": "^2.8.17",
+    "@types/express": "^5.0.0",
     "expect.js": "^0.3.1",
     "mocha": "^8.0.1",
     "nodemon": "^3.1.9",
