+++ backend/src/routes/index.js
@@ -1,0 +1,6 @@
+import jobRouter from "./job.js";
+import namespaceRouter from "./namespace.js";
+import podRouter from "./pod.js";
+import queueRouter from "./queue.js";
+
+export { jobRouter, namespaceRouter, podRouter, queueRouter };
