+++ backend/src/routes/namespace.js
@@ -1,0 +1,8 @@
+import express from "express";
+import { getNamespaces } from "../controllers/namespace.js";
+
+const router = express.Router();
+
+router.get("/namespaces", getNamespaces);
+
+export default router;
